+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix #6 - make tab say 'Destiny Raid Status' instead of 'DRS'.
 * Fix #8 - add link to Destiny Status and Destiny Tracker for each gamertag.
+* Semi-fix #3 - add some @media queries
 * Fixes to gamertag input (empty, whitespace).
 * Fix Chrome extension version number differing from website.
 * Add 'otherwise' route back to index.